Put the dashboard as the default layout
